refactor!: remove non-db totalizer encodings
- remove old `Totalizer` and `GeneralizedTotalizer`
- rename `DbTotalizer` -> `Totalizer`
- rename `DbGte` -> `GeneralizedTotalizer`
