+++ examples/profiling.rs
@@ -1,7 +1,7 @@
 //! # Small Executable For Profiling Tasks
 
 use rustsat::{
-    encodings::pb::{BoundUpper, DbGte},
+    encodings::pb::{BoundUpper, GeneralizedTotalizer},
     instances::{BasicVarManager, Cnf, ManageVars},
     lit,
     types::{Lit, Var},
@@ -78,7 +78,7 @@ fn build_full_ub<PBE: BoundUpper + FromIterator<(Lit, usize)>>(lits: &[(Lit, usi
 
 fn main() {
     for idx in 0..1 {
-        build_full_ub::<DbGte>(&lits!());
+        build_full_ub::<GeneralizedTotalizer>(&lits!());
         println!("iteration {}", idx);
     }
 }
