+++ tests/maxsat.rs
@@ -2,7 +2,8 @@ use rustsat::{algs::maxsat, encodings::pb, instances::OptInstance};
 
 #[test]
 fn sis_small() {
-    type Alg = maxsat::SolutionImprovingSearch<rustsat_minisat::core::Minisat, pb::DbGte>;
+    type Alg =
+        maxsat::SolutionImprovingSearch<rustsat_minisat::core::Minisat, pb::GeneralizedTotalizer>;
     let inst: OptInstance = OptInstance::from_dimacs_path("./data/inc-sis-fails.wcnf").unwrap();
     let sol = inst.solve_maxsat::<Alg>();
     dbg!(&sol);
@@ -11,7 +12,8 @@ fn sis_small() {
 
 #[test]
 fn sis() {
-    type Alg = maxsat::SolutionImprovingSearch<rustsat_minisat::core::Minisat, pb::DbGte>;
+    type Alg =
+        maxsat::SolutionImprovingSearch<rustsat_minisat::core::Minisat, pb::GeneralizedTotalizer>;
     let inst: OptInstance =
         OptInstance::from_dimacs_path("./data/auctions_wt-cat_sched_60_70_0003.txt.wcnf").unwrap();
     let sol = inst.solve_maxsat::<Alg>();
